app/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0